+++ src/models/mod.rs
@@ -1,5 +1,6 @@
 pub mod attachment;
 pub mod identifiers;
 pub mod incident;
+pub mod order_by;
 pub mod record;
 pub mod secret;